Consolidate theming storage
Shouldn't this go in `openedx/core/djangoapps/theming`, the subdirectory
created to house, well, OpenEdX's core Django Theming ;)
