+++ cms/envs/aws.py
@@ -176,6 +176,7 @@ COURSES_WITH_UNSAFE_CODE = ENV_TOKENS.get("COURSES_WITH_UNSAFE_CODE", [])
 ASSET_IGNORE_REGEX = ENV_TOKENS.get('ASSET_IGNORE_REGEX', ASSET_IGNORE_REGEX)
 
 # Theme overrides
+STATICFILES_STORAGE = 'openedx.core.djangoapps.theming.storage.ProductionStorage'
 THEME_NAME = ENV_TOKENS.get('THEME_NAME', None)
 COMPREHENSIVE_THEMING_DIRECTORY = path(
     ENV_TOKENS.get(
