+++ cms/envs/devstack.py
@@ -37,7 +37,7 @@ FEATURES['PREVIEW_LMS_BASE'] = "preview." + LMS_BASE
 
 # Skip packaging and optimization in development
 PIPELINE_ENABLED = False
-STATICFILES_STORAGE = 'openedx.core.storage.DevelopmentStorage'
+STATICFILES_STORAGE = 'openedx.core.djangoapps.theming.storage.DevelopmentStorage'
 
 # Revert to the default set of finders as we don't want the production pipeline
 STATICFILES_FINDERS = [
