+++ lms/envs/aws.py
@@ -236,6 +236,7 @@ BULK_EMAIL_ROUTING_KEY = HIGH_PRIORITY_QUEUE
 BULK_EMAIL_ROUTING_KEY_SMALL_JOBS = LOW_PRIORITY_QUEUE
 
 # Theme overrides
+STATICFILES_STORAGE = 'openedx.core.djangoapps.theming.storage.ProductionStorage'
 THEME_NAME = ENV_TOKENS.get('THEME_NAME', None)
 COMPREHENSIVE_THEMING_DIRECTORY = path(
     ENV_TOKENS.get(
