+++ lms/envs/devstack.py
@@ -87,7 +87,7 @@ def should_show_debug_toolbar(_):
 ########################### PIPELINE #################################
 
 PIPELINE_ENABLED = False
-STATICFILES_STORAGE = 'openedx.core.storage.DevelopmentStorage'
+STATICFILES_STORAGE = 'openedx.core.djangoapps.theming.storage.DevelopmentStorage'
 
 # Revert to the default set of finders as we don't want the production pipeline
 STATICFILES_FINDERS = [
